refactor: bazel test pkg rename
