+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/bazel/functional/bazel/pipeline/step/IntermediateStepSplitEachEachTest.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.bazel.functional.bazel.pipeline.stepexecutor;
+package com.synopsys.integration.detectable.detectables.bazel.functional.bazel.pipeline.step;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
